requirements.txt
flask
python-dotenv
mysql.connector
gunicorn